Change to not mark past stop times as unavailable
If a provided trip update does not start with the first stop time of a trip, all the previous stop times is marked as unavailble.

This change that behaviour to instead set the delay for these stop times to zero. This will allow these stop times to be available in routing.